# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yui VRM AI Studio Backend",
    "description": "Local BYOK backend for Yui VRM AI Studio.",
    "version": "0.2.0-beta.5"
  },
  "paths": {
...
    "/admin/api/settings": {
      "patch": {
        "summary": "Update Settings",
        "operationId": "update_settings_admin_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/events": {
      "get": {
        "summary": "Activity Events",
        "operationId": "activity_events_admin_api_events_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/probe/{provider_id}": {
      "post": {
        "summary": "Probe",
        "operationId": "probe_admin_api_probe__provider_id__post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/voice-library": {
      "get": {
        "summary": "Voice Library",
        "operationId": "voice_library_admin_api_voice_library_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Voice Library",
        "operationId": "update_voice_library_admin_api_voice_library_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/voice-endpoints/{endpoint_id}/probe": {
      "post": {
        "summary": "Probe Voice Endpoint",
        "operationId": "probe_voice_endpoint_admin_api_voice_endpoints__endpoint_id__probe_post",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/clear-scope": {
      "post": {
        "summary": "Clear Scope",
        "operationId": "clear_scope_admin_api_clear_scope_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearScope"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/voice-preview": {
      "post": {
        "summary": "Preview Voice",
        "operationId": "preview_voice_admin_api_voice_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoicePreview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ClearScope": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "include_memories": {
            "type": "boolean",
            "title": "Include Memories",
            "default": false
          },
          "confirmation": {
            "type": "string",
            "title": "Confirmation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user_id",
          "confirmation"
        ],
        "title": "ClearScope"
      },
...
      "Endpoint": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "provider_type": {
            "type": "string",
            "enum": [
              "voicevox",
              "aivis",
              "http"
            ],
            "title": "Provider Type"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 40,
            "title": "Settings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "provider_type"
        ],
        "title": "Endpoint"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryUpdate": {
        "properties": {
          "revision": {
            "type": "string",
            "maxLength": 128,
            "title": "Revision"
          },
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Endpoints"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/Profile"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Profiles"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "endpoints",
          "profiles"
        ],
        "title": "LibraryUpdate"
      },
...
      "Profile": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "endpoint_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Endpoint Id"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 12,
            "title": "Parameters"
          },
          "voice": {
            "type": "string",
            "maxLength": 200,
            "title": "Voice",
            "default": ""
          },
          "model": {
            "type": "string",
            "maxLength": 200,
            "title": "Model",
            "default": ""
          },
          "fallback_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[A-Za-z0-9_.-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Profile Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "endpoint_id"
        ],
        "title": "Profile"
      },
...
      "SettingsUpdate": {
        "properties": {
          "revision": {
            "type": "string",
            "maxLength": 128,
            "title": "Revision"
          },
          "changes": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 100,
            "title": "Changes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "changes"
        ],
        "title": "SettingsUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VoicePreview": {
        "properties": {
          "endpoint_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Endpoint Id"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 12,
            "title": "Parameters"
          },
          "voice": {
            "type": "string",
            "maxLength": 200,
            "title": "Voice",
            "default": ""
          },
          "model": {
            "type": "string",
            "maxLength": 200,
            "title": "Model",
            "default": ""
          },
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "endpoint_id",
          "text"
        ],
        "title": "VoicePreview"
      },
...
    }
  }
}